+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "clickhouse_gdc"
-version = "2.37.1"
+version = "2.37.2"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
