+++ src/server/routes/post_query.rs
@@ -9,7 +9,7 @@ use crate::{
         config::{SourceConfig, SourceName},
         error::ServerError,
     },
-    sql::{apply_aliases_to_query_request, QueryBuilder},
+    sql::QueryBuilder,
 };
 
 #[axum_macros::debug_handler]
@@ -18,8 +18,7 @@ pub async fn post_query(
     SourceConfig(config): SourceConfig,
     WithRejection(Json(request), _): WithRejection<Json<QueryRequest>, ServerError>,
 ) -> Result<Json<QueryResponse>, ServerError> {
-    let request = apply_aliases_to_query_request(request, &config)?;
-    let statement = QueryBuilder::build_sql_statement(&request, false)?;
+    let statement = QueryBuilder::build_sql_statement(&request, false, &config)?;
 
     let statement_string = statement.to_string();
