+++ src/sql.rs
@@ -1,5 +1,3 @@
 mod ast;
 mod query_builder;
-pub use query_builder::{
-    aliasing::apply_aliases_to_query_request, QueryBuilder, QueryBuilderError,
-};
+pub use query_builder::{QueryBuilder, QueryBuilderError};
